refactor: reorganized namespaces per layer
