+++ Rollbar/Infrastructure/InternalRollbarError.cs
@@ -1,4 +1,4 @@
-﻿namespace Rollbar
+﻿namespace Rollbar.Infrastructure
 {
     using System;
