+++ Rollbar/Infrastructure/RollbarQueueController.cs
@@ -1,6 +1,6 @@
 ﻿#define TRACE
 
-namespace Rollbar
+namespace Rollbar.Infrastructure
 {
     using Rollbar.Common;
     using Rollbar.Diagnostics;
