+++ Rollbar/Infrastructure/_RollbarClients.cd
@@ -1,0 +1,2 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ClassDiagram />
