+++ Rollbar/RollbarInternalEvents/PayloadDropEventArgs.cs
@@ -1,9 +1,10 @@
 ﻿namespace Rollbar
 {
-    using Rollbar.DTOs;
     using System;
-    using System.Collections.Generic;
     using System.Text;
+
+    using Rollbar.DTOs;
+    using Rollbar.Infrastructure;
 
     /// <summary>
     /// Class PayloadDropEventArgs.
