+++ UnitTest.Rollbar/AccessTokenQueuesMetadataFixture.cs
@@ -1,6 +1,8 @@
 ﻿namespace UnitTest.Rollbar
 {
     using global::Rollbar;
+    using global::Rollbar.Infrastructure;
+
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using System;
